src/utils/config.py
import yaml
from pathlib import Path
from typing import Any, Dict

def load_config(path: str = "configs/default.yaml") -> Dict[str, Any]:
    with open(path, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)

def ensure_dirs(*paths: str) -> None:
    for p in paths:
        Path(p).mkdir(parents=True, exist_ok=True)
